Perform deletion check prior to deleting file

+++ setup/setup.sh
@@ -62,14 +62,20 @@ case ${osinfo} in
       tar -xvf geckodriver-v0.22.0-linux64.tar.gz
       rm geckodriver-v0.22.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     else
       wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux32.tar.gz
       tar -xvf geckodriver-v0.22.0-linux32.tar.gz
       rm geckodriver-v0.22.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     fi
     cd ..
@@ -151,14 +157,20 @@ case ${osinfo} in
       tar -xvf geckodriver-v0.22.0-linux64.tar.gz
       rm geckodriver-v0.22.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     else
       wget https://github.com/mozilla/geckodriver/releases/download/v0.22.0/geckodriver-v0.22.0-linux32.tar.gz
       tar -xvf geckodriver-v0.22.0-linux32.tar.gz
       rm geckodriver-v0.22.0-linux64.tar.gz
       mv geckodriver /usr/sbin
-      rm /usr/bin/geckodriver
+      if [ -e /usr/bin/geckodriver ]
+      then
+      	rm /usr/bin/geckodriver
+      fi
       ln -s /usr/sbin/geckodriver /usr/bin/geckodriver
     fi
     cd ..
